All code moved from header files
